Better seperation of files; get rid of custom libs; improvments
