+++ imports/ui/App.tsx
@@ -9,7 +9,7 @@ import TaskList from "./components/TaskList"
 import {Task} from "./components/Task"
 import {ZeebeStatus} from "/imports/ui/components/ZeebeStatus"
 import {Toaster} from "react-hot-toast";
-import {DoctoralSchoolsForm} from "/imports/ui/components/DoctoralSchoolsForm";
+import {DoctoralSchoolsForm} from "/imports/ui/components/DoctoralSchools/form";
 import {Dashboard} from "/imports/ui/components/Dashboard";
 import {useTracker} from "meteor/react-meteor-data";
 import {Meteor} from "meteor/meteor";
