+++ package.json
@@ -30,7 +30,6 @@
     "plotly.js": "^2.6.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
-    "react-easy-edit": "^1.15.0",
     "react-hot-toast": "^2.0.0",
     "react-popover": "^0.5.10",
     "react-popper": "^2.2.4",
